requirements.txt
opencv-python~=4.5.2.54
numpy~=1.20.3
requests~=2.25.1
python-firebase~=1.2